Move user-login mock to correct location.
