+++ src/helpers/shared/__mocks__/user-login.js
@@ -1,8 +1,5 @@
 import MockAdapter from 'axios-mock-adapter';
-import {
-  getAxiosInstance,
-  getGraphqlInstance
-} from '../../helpers/shared/user-login';
+import { getAxiosInstance, getGraphqlInstance } from '../user-login';
 
 export const mockApi = new MockAdapter(getAxiosInstance());
 export const mockGraphql = new MockAdapter(getGraphqlInstance());
