+++ src/test/.eslintrc
@@ -1,7 +1,8 @@
 {
   "env": {
     "jest": true,
-    "es6": true
+    "es6": true,
+    "jasmine": true
   },
   "globals": {
     "apiClientMock": true,
