+++ src/test/redux/actions/approval-actions.test.js
@@ -5,7 +5,7 @@ import { APPROVAL_API_BASE } from '../../../utilities/constants';
 import { fetchWorkflows } from '../../../redux/actions/approval-actions';
 import { ASYNC_ACTIONS } from '../../../redux/action-types/approval-action-types';
 import { notificationsMiddleware } from '@redhat-cloud-services/frontend-components-notifications/';
-import { mockApi } from '../../__mocks__/user-login';
+import { mockApi } from '../../../helpers/shared/__mocks__/user-login';
 
 describe('approval actions', () => {
   const middlewares = [thunk, promiseMiddleware, notificationsMiddleware()];
