+++ src/test/redux/actions/order-actions.test.js
@@ -18,7 +18,7 @@ import {
   SET_SELECTED_PLAN,
   SUBMIT_SERVICE_ORDER
 } from '../../../redux/action-types';
-import { mockApi } from '../../__mocks__/user-login';
+import { mockApi } from '../../../helpers/shared/__mocks__/user-login';
 
 describe('Order actions', () => {
   const middlewares = [thunk, promiseMiddleware, notificationsMiddleware()];
