+++ src/test/smart-components/portfolio/portfolio-item-detail/edit-portfolio-item.test.js
@@ -17,7 +17,7 @@ import {
   UPDATE_PORTFOLIO_ITEM
 } from '../../../../redux/action-types';
 import { openApiReducerMock } from '../../../__mocks__/open-api-mock';
-import { mockApi } from '../../../__mocks__/user-login';
+import { mockApi } from '../../../../helpers/shared/__mocks__/user-login';
 
 describe('<EditPortfolioItem />', () => {
   const middlewares = [thunk, promiseMiddleware, notificationsMiddleware()];
