+++ src/test/smart-components/portfolio/portfolio-item-detail/portfolio-item-details.test.js
@@ -19,9 +19,9 @@ import {
 } from '../../../../utilities/constants';
 import ItemDetailDescription from '../../../../smart-components/portfolio/portfolio-item-detail/item-detail-description';
 import { PortfolioItemDetailToolbar } from '../../../../smart-components/portfolio/portfolio-item-detail/portfolio-item-detail-toolbar';
-import { mockApi } from '../../../__mocks__/user-login';
 import { Alert } from '@patternfly/react-core';
 import UserContext from '../../../../user-context';
+import { mockApi } from '../../../../helpers/shared/__mocks__/user-login';
 
 describe('<PortfolioItemDetail />', () => {
   let initialProps;
