+++ src/test/smart-components/portfolio/remove-portfolio-modal.test.js
@@ -19,7 +19,7 @@ import {
   FETCH_PORTFOLIOS,
   DELETE_TEMPORARY_PORTFOLIO
 } from '../../../redux/action-types';
-import { mockApi } from '../../__mocks__/user-login';
+import { mockApi } from '../../../helpers/shared/__mocks__/user-login';
 
 describe('<RemovePortfolioModal />', () => {
   let initialProps;
